+++ maintenance/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class MaintenanceConfig(AppConfig):
+    name = "maintenance"